mobile/appium-python/requirements.txt
Appium-Python-Client==3.1.1
pytest==7.4.3
python-dotenv==1.0.0